# Replace `cloc` with `count-loc`, using much faster SCC program