The `cloc` Perl script is not great for performance. There are several modern replacements, including Tokei (Rust), Loc (Rust), and SCC (Go). Loc has recommended that people now use SCC. Tokei looks promising, but SCC has better benchmarks https://boyter.org/posts/sloc-cloc-code-performance-update/. SCC is very well maintained with several maintainers https://github.com/boyter/scc, and they have several blogs on how they achieved such impressive performance https://boyter.org/posts/sloc-cloc-code-performance/. 

This also allows us to simplify our implementation. We are now nothing more than a light wrapper around SCC, in line with our general philosophy with Pants 2.0 and the v2 engine. We expose a new option `--args` which allows users to customize the tool as they want, e.g. `./pants count-loc '**' -- --no-complexity -v`.

Sample run on Pants:

```

───────────────────────────────────────────────────────────────────────────────
Language                 Files     Lines   Blanks  Comments     Code Complexity
───────────────────────────────────────────────────────────────────────────────
Python                     526     77465     6835      5246    65384       3197
Rust                       105     37940     3609      3394    30937       1392
Markdown                    34      1457      287         0     1170          0
ReStructuredText            34     18759     6164         0    12595          0
TOML                        33       921       76        88      757          0
Shell                       29      1645      272       250     1123        159
Protocol Buffers            21      4970      517      3400     1053          0
Plain Text                   7       119        7         0      112          0
BASH                         6       468       83       129      256         60
Dockerfile                   4       205       34        42      129          4
License                      4       646      104         0      542          0
gitignore                    4        75        0         6       69          0
Mustache                     2        55        9         0       46          0
YAML                         2      1006        6        11      989          0
C                            1         0        0         0        0          0
Cargo Lock                   1      3801      367         2     3432          0
XML                          1         7        0         0        7          0
───────────────────────────────────────────────────────────────────────────────
Total                      814    149539    18370     12568   118601       4812
───────────────────────────────────────────────────────────────────────────────
Estimated Cost to Develop $4,068,066
Estimated Schedule Effort 26.149191 months
Estimated People Required 18.428267
───────────────────────────────────────────────────────────────────────────────
```

### Benchmark on Pants codebase
Both benchmarks use our UUID hack to force rerunning the Cloc process, but use the cache for downloading the program.

Before:

```
1: ./pants --no-pantsd cloc '**'
            Mean        Std.Dev.    Min         Median      Max
real        3.560       0.045       3.492       3.568       3.650
user        3.337       0.029       3.287       3.336       3.386
sys         1.258       0.033       1.199       1.263       1.292
```

After:

```
1: ./pants --no-pantsd count-loc '**'
            Mean        Std.Dev.    Min         Median      Max
real        1.949       0.069       1.838       1.951       2.094
user        1.833       0.032       1.797       1.828       1.910
sys         1.136       0.026       1.088       1.134       1.177
```

Almost all of the time is now from Pants.

[ci skip-rust]
[ci skip-build-wheels]